feat: add script to print Star Wars movie title by episode number
This commit adds a script that prints the title of a Star Wars movie based on the provided episode number. The script takes the movie ID as the first argument and utilizes the Star Wars API endpoint https://swapi-api.alx-tools.com/api/films/:id to fetch the movie details. It then extracts and prints the title of the movie corresponding to the given episode number.